Job_Hunter/requirements.txt
Django==3.2.8
django-background-task==0.1.8
django-cors-headers==3.10.1
djangorestframework==3.12.4
beautifulsoup4==4.10.0
requests==2.22.0